backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
firebase-admin==6.5.0
reportlab==4.2.2
twilio==9.3.3
python-dotenv==1.0.1
pydantic==2.8.2
